feat: Android AI Activation Full Feature Optimization

+++ PoseAI-Android/app/src/main/java/com/poseai/app/data/Entities.kt
@@ -69,7 +69,7 @@ data class SceneStat(
 )
 
 data class ScoreStat(
-    val avgScore: Float,
+    val avgScore: Double,
     val maxScore: Int,
     val count: Int
 )
